+++ webapp/src/views/Decks.vue
@@ -1,14 +1,13 @@
 <template>
   <div>
-    <div
-      v-show="warning"
-      @click="closeOverlay()"
-      class="overlay"
-    ></div>
+    <div v-show="warning" @click="closeOverlay()" class="overlay"></div>
 
     <span v-if="previewCard">
       <div class="card-preview">
-        <img :src="`/assets/cards/all/${previewCard.uid}.jpg`" alt="Full card">
+        <img
+          :src="`/assets/cards/all/${previewCard.uid}.jpg`"
+          alt="Full card"
+        />
       </div>
     </span>
 
@@ -143,7 +142,9 @@
               <td
                 @mouseover="previewCard = card"
                 @mouseleave="previewCard = null"
-              >{{ card.name }}</td>
+              >
+                {{ card.name }}
+              </td>
               <td class="set">{{ card.set }}</td>
               <td class="civilization">{{ card.civilization }}</td>
               <td class="manaCost">{{ card.manaCost }}</td>
@@ -204,7 +205,9 @@
               src="/assets/images/delete_icon.png"
             />
           </a>
-          <div @click="copyDeckList()" class="btn copy-deck-list">Copy Deck List</div>
+          <div @click="copyDeckList()" class="btn copy-deck-list">
+            Copy Deck List
+          </div>
           <div @click="newDeck()" class="btn new">New Deck</div>
           <template
             v-if="
@@ -277,7 +280,7 @@
                     />
                   </div>
                 </th>
-                
+
                 <th>
                   <div class="sort-btn" @click="toggleSortForDeck('family')">
                     Race
@@ -309,7 +312,9 @@
                   <td
                     @mouseover="previewCard = card"
                     @mouseleave="previewCard = null"
-                  >{{ card.name }}</td>
+                  >
+                    {{ card.name }}
+                  </td>
                   <td class="set">{{ card.set }}</td>
                   <td class="civilization">{{ card.civilization }}</td>
                   <td class="manaCost">{{ card.manaCost }}</td>
@@ -361,7 +366,7 @@ export default {
       },
       sortForDeck: {
         by: "name",
-        directionNum: 1,
+        directionNum: 1
       },
 
       sets: [],
@@ -399,7 +404,7 @@ export default {
       this.sortForDeck = {
         directionNum:
           this.sortForDeck.by === by ? -this.sortForDeck.directionNum : 1,
-        by,
+        by
       };
     },
 
@@ -419,11 +424,17 @@ export default {
         }
       }
 
-      cards.sort(
-        (c1, c2) =>
-          (c1[this.sortForDeck.by] === parseInt(c1[this.sortForDeck.by], 10) && c2[this.sortForDeck.by] === parseInt(c2[this.sortForDeck.by], 10)) ? 
-            this.sortForDeck.directionNum * (c1[this.sortForDeck.by] < c2[this.sortForDeck.by] ? -1 : c1[this.sortForDeck.by] > c2[this.sortForDeck.by] ? 1 : 0) : 
-            this.sortForDeck.directionNum * c1[this.sortForDeck.by].localeCompare(c2[this.sortForDeck.by])
+      cards.sort((c1, c2) =>
+        c1[this.sortForDeck.by] === parseInt(c1[this.sortForDeck.by], 10) &&
+        c2[this.sortForDeck.by] === parseInt(c2[this.sortForDeck.by], 10)
+          ? this.sortForDeck.directionNum *
+            (c1[this.sortForDeck.by] < c2[this.sortForDeck.by]
+              ? -1
+              : c1[this.sortForDeck.by] > c2[this.sortForDeck.by]
+              ? 1
+              : 0)
+          : this.sortForDeck.directionNum *
+            c1[this.sortForDeck.by].localeCompare(c2[this.sortForDeck.by])
       );
 
       return cards;
@@ -467,7 +478,9 @@ export default {
       const cards = this.getCardsForDeck(this.selectedDeck.cards);
 
       const deckList = cards.map(card => `${card.count}x ${card.name}`);
-      this.warning = `${deckList.join("\n")}\n\nTotal Cards: ${this.selectedDeck.cards.length}`;
+      this.warning = `${deckList.join("\n")}\n\nTotal Cards: ${
+        this.selectedDeck.cards.length
+      }`;
     },
 
     newDeck() {
@@ -641,11 +654,17 @@ export default {
         );
       }
 
-      cards.sort(
-        (c1, c2) =>
-          (c1[this.sort.by] === parseInt(c1[this.sort.by], 10) && c2[this.sort.by] === parseInt(c2[this.sort.by], 10)) ? 
-            this.sort.directionNum * (c1[this.sort.by] < c2[this.sort.by] ? -1 : c1[this.sort.by] > c2[this.sort.by] ? 1 : 0) : 
-            this.sort.directionNum * c1[this.sort.by].localeCompare(c2[this.sort.by])
+      cards.sort((c1, c2) =>
+        c1[this.sort.by] === parseInt(c1[this.sort.by], 10) &&
+        c2[this.sort.by] === parseInt(c2[this.sort.by], 10)
+          ? this.sort.directionNum *
+            (c1[this.sort.by] < c2[this.sort.by]
+              ? -1
+              : c1[this.sort.by] > c2[this.sort.by]
+              ? 1
+              : 0)
+          : this.sort.directionNum *
+            c1[this.sort.by].localeCompare(c2[this.sort.by])
       );
 
       return cards;
@@ -673,15 +692,15 @@ export default {
         set: "arrow_up_down",
         civilization: "arrow_up_down",
         manaCost: "arrow_up_down",
-        family: "arrow_up_down",
+        family: "arrow_up_down"
       };
 
       result[this.sortForDeck.by] =
         this.sortForDeck.directionNum === 1 ? "arrow_down" : "arrow_up";
 
       return result;
-    },
-  },
+    }
+  }
 };
 </script>
 
@@ -767,13 +786,13 @@ export default {
 }
 
 .copy-deck-list {
-  background: #DC758F !important;
+  background: #dc758f !important;
   display: inline-block;
   margin-right: 15px;
 }
 
 .copy-deck-list:hover {
-  background: #D65C7A !important;
+  background: #d65c7a !important;
 }
 
 .discard {
@@ -867,44 +886,6 @@ input:active,
 textarea:active,
 select:active {
   outline: none;
-}
-
-.left::-webkit-scrollbar-track {
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  border-radius: 10px;
-  background-color: #484c52;
-}
-
-.left::-webkit-scrollbar {
-  width: 6px;
-  background-color: #484c52;
-}
-
-.left::-webkit-scrollbar-thumb {
-  border-radius: 10px;
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  background-color: #222;
-}
-
-.right::-webkit-scrollbar-track {
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  border-radius: 10px;
-  background-color: #484c52;
-}
-
-.right::-webkit-scrollbar {
-  width: 6px;
-  background-color: #484c52;
-}
-
-.right::-webkit-scrollbar-thumb {
-  border-radius: 10px;
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-  background-color: #222;
 }
 
 table {